extract common tests from span and log filtering
