--- internal/processor/filtermatcher/.nocover
@@ -1,1 +1,0 @@
-Tested in filterspan package
