+++ internal/processor/filterspan/filterspan_test.go
@@ -34,7 +34,6 @@ func createConfig(matchType filterset.MatchType) *filterset.Config {
 }
 
 func TestSpan_validateMatchesConfiguration_InvalidConfig(t *testing.T) {
-	version := "["
 	testcases := []struct {
 		name        string
 		property    filterconfig.MatchProperties
@@ -75,26 +74,6 @@ func TestSpan_validateMatchesConfiguration_InvalidConfig(t *testing.T) {
 			errorString: "error creating service name filters: unrecognized match_type: '', valid types are: [regexp strict]",
 		},
 		{
-			name: "regexp_match_type_for_int_attribute",
-			property: filterconfig.MatchProperties{
-				Config: *createConfig(filterset.Regexp),
-				Attributes: []filterconfig.Attribute{
-					{Key: "key", Value: 1},
-				},
-			},
-			errorString: `error creating attribute filters: match_type=regexp for "key" only supports STRING, but found INT`,
-		},
-		{
-			name: "unknown_attribute_value",
-			property: filterconfig.MatchProperties{
-				Config: *createConfig(filterset.Strict),
-				Attributes: []filterconfig.Attribute{
-					{Key: "key", Value: []string{}},
-				},
-			},
-			errorString: `error creating attribute filters: error unsupported value type "[]string"`,
-		},
-		{
 			name: "invalid_regexp_pattern_service",
 			property: filterconfig.MatchProperties{
 				Config:   *createConfig(filterset.Regexp),
@@ -110,52 +89,6 @@ func TestSpan_validateMatchesConfiguration_InvalidConfig(t *testing.T) {
 			},
 			errorString: "error creating span name filters: error parsing regexp: missing closing ]: `[`",
 		},
-		{
-			name: "invalid_regexp_pattern_attribute",
-			property: filterconfig.MatchProperties{
-				Config:     *createConfig(filterset.Regexp),
-				SpanNames:  []string{"["},
-				Attributes: []filterconfig.Attribute{{Key: "key", Value: "["}},
-			},
-			errorString: "error creating attribute filters: error parsing regexp: missing closing ]: `[`",
-		},
-		{
-			name: "invalid_regexp_pattern_resource",
-			property: filterconfig.MatchProperties{
-				Config:    *createConfig(filterset.Regexp),
-				Resources: []filterconfig.Attribute{{Key: "key", Value: "["}},
-			},
-			errorString: "error creating resource filters: error parsing regexp: missing closing ]: `[`",
-		},
-		{
-			name: "invalid_regexp_pattern_library_name",
-			property: filterconfig.MatchProperties{
-				Config:    *createConfig(filterset.Regexp),
-				Libraries: []filterconfig.InstrumentationLibrary{{Name: "["}},
-			},
-			errorString: "error creating library name filters: error parsing regexp: missing closing ]: `[`",
-		},
-		{
-			name: "invalid_regexp_pattern_library_version",
-			property: filterconfig.MatchProperties{
-				Config:    *createConfig(filterset.Regexp),
-				Libraries: []filterconfig.InstrumentationLibrary{{Name: "lib", Version: &version}},
-			},
-			errorString: "error creating library version filters: error parsing regexp: missing closing ]: `[`",
-		},
-		{
-			name: "empty_key_name_in_attributes_list",
-			property: filterconfig.MatchProperties{
-				Config:   *createConfig(filterset.Strict),
-				Services: []string{"a"},
-				Attributes: []filterconfig.Attribute{
-					{
-						Key: "",
-					},
-				},
-			},
-			errorString: "error creating attribute filters: can't have empty key in the list of attributes",
-		},
 	}
 	for _, tc := range testcases {
 		t.Run(tc.name, func(t *testing.T) {
@@ -167,7 +100,6 @@ func TestSpan_validateMatchesConfiguration_InvalidConfig(t *testing.T) {
 }
 
 func TestSpan_Matching_False(t *testing.T) {
-	version := "wrong"
 	testcases := []struct {
 		name       string
 		properties *filterconfig.MatchProperties
@@ -211,103 +143,13 @@ func TestSpan_Matching_False(t *testing.T) {
 				Attributes: []filterconfig.Attribute{},
 			},
 		},
-
-		{
-			name: "wrong_library_name",
-			properties: &filterconfig.MatchProperties{
-				Config:    *createConfig(filterset.Strict),
-				Services:  []string{},
-				Libraries: []filterconfig.InstrumentationLibrary{{Name: "wrong"}},
-			},
-		},
-		{
-			name: "wrong_library_version",
-			properties: &filterconfig.MatchProperties{
-				Config:    *createConfig(filterset.Strict),
-				Services:  []string{},
-				Libraries: []filterconfig.InstrumentationLibrary{{Name: "lib", Version: &version}},
-			},
-		},
-
-		{
-			name: "wrong_attribute_value",
-			properties: &filterconfig.MatchProperties{
-				Config:   *createConfig(filterset.Strict),
-				Services: []string{},
-				Attributes: []filterconfig.Attribute{
-					{
-						Key:   "keyInt",
-						Value: 1234,
-					},
-				},
-			},
-		},
-		{
-			name: "wrong_resource_value",
-			properties: &filterconfig.MatchProperties{
-				Config:   *createConfig(filterset.Strict),
-				Services: []string{},
-				Resources: []filterconfig.Attribute{
-					{
-						Key:   "keyInt",
-						Value: 1234,
-					},
-				},
-			},
-		},
-		{
-			name: "incompatible_attribute_value",
-			properties: &filterconfig.MatchProperties{
-				Config:   *createConfig(filterset.Strict),
-				Services: []string{},
-				Attributes: []filterconfig.Attribute{
-					{
-						Key:   "keyInt",
-						Value: "123",
-					},
-				},
-			},
-		},
-		{
-			name: "unsupported_attribute_value",
-			properties: &filterconfig.MatchProperties{
-				Config:   *createConfig(filterset.Regexp),
-				Services: []string{},
-				Attributes: []filterconfig.Attribute{
-					{
-						Key:   "keyMap",
-						Value: "123",
-					},
-				},
-			},
-		},
-		{
-			name: "property_key_does_not_exist",
-			properties: &filterconfig.MatchProperties{
-				Config:   *createConfig(filterset.Strict),
-				Services: []string{},
-				Attributes: []filterconfig.Attribute{
-					{
-						Key:   "doesnotexist",
-						Value: nil,
-					},
-				},
-			},
-		},
 	}
 
 	span := pdata.NewSpan()
 	span.InitEmpty()
 	span.SetName("spanName")
-	span.Attributes().InitFromMap(map[string]pdata.AttributeValue{
-		"keyInt": pdata.NewAttributeValueInt(123),
-		"keyMap": pdata.NewAttributeValueMap(),
-	})
-
 	library := pdata.NewInstrumentationLibrary()
-	library.InitEmpty()
-	library.SetName("lib")
-	library.SetVersion("ver")
+	resource := pdata.NewResource()
 
 	for _, tc := range testcases {
 		t.Run(tc.name, func(t *testing.T) {
@@ -315,37 +157,9 @@ func TestSpan_Matching_False(t *testing.T) {
 			require.NoError(t, err)
 			assert.NotNil(t, matcher)
 
-			assert.False(t, matcher.MatchSpan(span, resource("wrongSvc"), library))
+			assert.False(t, matcher.MatchSpan(span, resource, library))
 		})
 	}
-}
-
-func resource(service string) pdata.Resource {
-	r := pdata.NewResource()
-	r.InitEmpty()
-	r.Attributes().InitFromMap(map[string]pdata.AttributeValue{conventions.AttributeServiceName: pdata.NewAttributeValueString(service)})
-	return r
-}
-
-func TestSpan_MatchingCornerCases(t *testing.T) {
-	cfg := &filterconfig.MatchProperties{
-		Config:   *createConfig(filterset.Strict),
-		Services: []string{"svcA"},
-		Attributes: []filterconfig.Attribute{
-			{
-				Key:   "keyOne",
-				Value: nil,
-			},
-		},
-	}
-
-	mp, err := NewMatcher(cfg)
-	assert.Nil(t, err)
-	assert.NotNil(t, mp)
-
-	emptySpan := pdata.NewSpan()
-	emptySpan.InitEmpty()
-	assert.False(t, mp.MatchSpan(emptySpan, resource("svcA"), pdata.NewInstrumentationLibrary()))
 }
 
 func TestSpan_MissingServiceName(t *testing.T) {
@@ -360,12 +174,10 @@ func TestSpan_MissingServiceName(t *testing.T) {
 
 	emptySpan := pdata.NewSpan()
 	emptySpan.InitEmpty()
-	assert.False(t, mp.MatchSpan(emptySpan, resource(""), pdata.NewInstrumentationLibrary()))
+	assert.False(t, mp.MatchSpan(emptySpan, pdata.NewResource(), pdata.NewInstrumentationLibrary()))
 }
 
 func TestSpan_Matching_True(t *testing.T) {
-	ver := "v.*"
-
 	testcases := []struct {
 		name       string
 		properties *filterconfig.MatchProperties
@@ -383,22 +195,6 @@ func TestSpan_Matching_True(t *testing.T) {
 			properties: &filterconfig.MatchProperties{
 				Config:     *createConfig(filterset.Strict),
 				Services:   []string{"svcA"},
-				Attributes: []filterconfig.Attribute{},
-			},
-		},
-		{
-			name: "library_match",
-			properties: &filterconfig.MatchProperties{
-				Config:     *createConfig(filterset.Regexp),
-				Libraries:  []filterconfig.InstrumentationLibrary{{Name: "li.*"}},
-				Attributes: []filterconfig.Attribute{},
-			},
-		},
-		{
-			name: "library_match_with_version",
-			properties: &filterconfig.MatchProperties{
-				Config:     *createConfig(filterset.Regexp),
-				Libraries:  []filterconfig.InstrumentationLibrary{{Name: "li.*", Version: &ver}},
 				Attributes: []filterconfig.Attribute{},
 			},
 		},
@@ -423,97 +219,6 @@ func TestSpan_Matching_True(t *testing.T) {
 				Attributes: []filterconfig.Attribute{},
 			},
 		},
-		{
-			name: "attribute_exact_value_match",
-			properties: &filterconfig.MatchProperties{
-				Config:   *createConfig(filterset.Strict),
-				Services: []string{},
-				Attributes: []filterconfig.Attribute{
-					{
-						Key:   "keyString",
-						Value: "arithmetic",
-					},
-					{
-						Key:   "keyInt",
-						Value: 123,
-					},
-					{
-						Key:   "keyDouble",
-						Value: 3245.6,
-					},
-					{
-						Key:   "keyBool",
-						Value: true,
-					},
-				},
-			},
-		},
-		{
-			name: "attribute_regex_value_match",
-			properties: &filterconfig.MatchProperties{
-				Config: *createConfig(filterset.Regexp),
-				Attributes: []filterconfig.Attribute{
-					{
-						Key:   "keyString",
-						Value: "arith.*",
-					},
-					{
-						Key:   "keyInt",
-						Value: "12.*",
-					},
-					{
-						Key:   "keyDouble",
-						Value: "324.*",
-					},
-					{
-						Key:   "keyBool",
-						Value: "tr.*",
-					},
-				},
-			},
-		},
-		{
-			name: "resource_exact_value_match",
-			properties: &filterconfig.MatchProperties{
-				Config: *createConfig(filterset.Strict),
-				Resources: []filterconfig.Attribute{
-					{
-						Key:   "resString",
-						Value: "arithmetic",
-					},
-				},
-			},
-		},
-		{
-			name: "property_exists",
-			properties: &filterconfig.MatchProperties{
-				Config:   *createConfig(filterset.Strict),
-				Services: []string{"svcA"},
-				Attributes: []filterconfig.Attribute{
-					{
-						Key:   "keyExists",
-						Value: nil,
-					},
-				},
-			},
-		},
-		{
-			name: "match_all_settings_exists",
-			properties: &filterconfig.MatchProperties{
-				Config:   *createConfig(filterset.Strict),
-				Services: []string{"svcA"},
-				Attributes: []filterconfig.Attribute{
-					{
-						Key:   "keyExists",
-						Value: nil,
-					},
-					{
-						Key:   "keyString",
-						Value: "arithmetic",
-					},
-				},
-			},
-		},
 	}
 
 	span := pdata.NewSpan()
@@ -532,13 +237,9 @@ func TestSpan_Matching_True(t *testing.T) {
 	resource.InitEmpty()
 	resource.Attributes().InitFromMap(map[string]pdata.AttributeValue{
 		conventions.AttributeServiceName: pdata.NewAttributeValueString("svcA"),
-		"resString":                      pdata.NewAttributeValueString("arithmetic"),
 	})
 
 	library := pdata.NewInstrumentationLibrary()
-	library.InitEmpty()
-	library.SetName("lib")
-	library.SetVersion("ver")
 
 	for _, tc := range testcases {
 		t.Run(tc.name, func(t *testing.T) {
